Add a title to 'dedication' environment and format more like the Guidelines.
This commit also eliminates the 'quotation' environment from being used in the dedication environment.  If one wants their dedication to be put in 'quotation', I believe they should do it manually.  To obtain what's shown in the Guidelines document supplied by the University, this and the initial \vfill needs to be removed.